% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bx_download}
\alias{bx_download}
\title{Download PDF's}
\usage{
bx_download(bxso, directory, create = TRUE)
}
\arguments{
\item{bxso}{search results from bx_search()}

\item{directory}{The location you want to download the PDF's to.}

\item{create}{if true create the directory if it doesn't exist}
}
\description{
Download PDF's of all the papers in your search results.
}
\examples{
\dontrun{
bxEco <- bx_search("ecology",limit=20)
bx_download(bxEco,"~/biorxivPDF)
}
}

