% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bx_extract_single}
\alias{bx_extract_single}
\title{Extract data from a single record}
\usage{
bx_extract_single(bxso_url)
}
\value{
a single S3 object with all the salient details about a paper
}
\description{
Generate an S3 object that represents a single paper
}

