% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bx_extract.R
\name{plot.biorxiv_paper}
\alias{plot.biorxiv_paper}
\title{plot metric details for a paper}
\usage{
\method{plot}{biorxiv_paper}(x, type = "abs", ...)
}
\arguments{
\item{x}{the paper to plot a summary of}

\item{type}{the data to plot, 'abs' for abstract views, 'dl' for PDF downloads}

\item{...}{extra parameters to pass}
}
\description{
plot a summary of the views a paper has had
}

