% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bx_search.R
\name{print.summary.biorxiv_search}
\alias{print.summary.biorxiv_search}
\title{Print summary results}
\usage{
\method{print}{summary.biorxiv_search}(x, ...)
}
\arguments{
\item{x}{the biorxiv search object to print}

\item{...}{extra parameters to the print function}
}
\description{
print a summary of search results
}

