% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bx_search.R
\name{summary.biorxiv_search}
\alias{summary.biorxiv_search}
\title{Summary of search results}
\usage{
\method{summary}{biorxiv_search}(object, ...)
}
\arguments{
\item{object}{the search object to create a summary of}

\item{...}{extra parameters to pass}
}
\description{
create a summary of search results
}

