% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bx_extract.R
\name{bx_extract}
\alias{bx_extract}
\title{Extract data from search results}
\usage{
bx_extract(bxso)
}
\arguments{
\item{bxso}{search results from bx_search()}
}
\value{
a list holding the details about all the papers from search results. Each element in the list is an S3 object that holds all the salient details about the paper
}
\description{
Extract information about each paper in search results, returning author info, abstract and download metrics
}
\details{
The metrics field will return a month by month data frame with abstract views and pdf downloads
}
\examples{
\dontrun{
bxEco <- bx_search("ecology",limit=5)
bxEcoData <- bx_extract(bxEco)
## See a plot of abstract views
plot(bxEcoData[[1]],type="abs")

## See a plot of downloads
plot(bxEcoData[[1]],type="dl)

}
}

