% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bx_extract.R
\name{bx_extract_single}
\alias{bx_extract_single}
\title{Extract data from a single record}
\usage{
bx_extract_single(bxso_url)
}
\arguments{
\item{bxso_url}{the URL of a single biorxiv paper}
}
\value{
a single S3 object with all the salient details about a paper
}
\description{
Generate an S3 object that represents a single paper
}

